Split core types into multiple files
